refactor(proto-common): decouple from datafusion-common
datafusion-proto-common is now standalone (no datafusion-* deps): it
owns the prost-generated types plus the foreign-foreign Arrow conv
impls (DataType, Field, Schema, TimeUnit, IntervalUnit, QuoteStyle).
Conversions whose target lives in datafusion-common now live in
`datafusion_common::proto` (gated behind a new `proto` feature).

The proto-common error types lose their `DataFusionError` variant,
gaining instead an `Arrow(arrow::error::ArrowError)` variant on
`FromProtoError`. `From<FromProtoError> for DataFusionError` and
`From<ToProtoError> for DataFusionError` live in datafusion-common.

This breaks the proto-common -> datafusion-common cycle and makes the
ownership rule unambiguous: each conv impl lives in the crate that owns
its target type.

datafusion-proto's own from_proto / to_proto modules now use
`DataFusionError` as the error type, since both proto-common and
datafusion-common errors converge there.
