+++ datafusion/datasource/Cargo.toml
@@ -34,7 +34,7 @@ all-features = true
 backtrace = ["datafusion-common/backtrace"]
 compression = ["async-compression", "liblzma", "bzip2", "flate2", "zstd", "tokio-util"]
 default = ["compression"]
-proto = ["dep:datafusion-proto-common"]
+proto = ["dep:datafusion-proto-common", "datafusion-common/proto"]
 
 [dependencies]
 arrow = { workspace = true }
