+++ datafusion/proto-common/Cargo.toml
@@ -17,7 +17,7 @@
 
 [package]
 name = "datafusion-proto-common"
-description = "Protobuf serialization of DataFusion common types"
+description = "Protobuf model types for DataFusion (prost-generated, no datafusion-* deps)"
 keywords = ["arrow", "query", "sql"]
 readme = "README.md"
 version = { workspace = true }
@@ -40,7 +40,6 @@ json = ["serde", "pbjson"]
 
 [dependencies]
 arrow = { workspace = true }
-datafusion-common = { workspace = true }
 pbjson = { workspace = true, optional = true }
 prost = { workspace = true }
 serde = { version = "1.0", optional = true }
