+++ datafusion/proto-common/src/common.rs
@@ -15,9 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use datafusion_common::{DataFusionError, internal_datafusion_err};
-
-/// Return a `DataFusionError::Internal` with the given message
-pub fn proto_error<S: Into<String>>(message: S) -> DataFusionError {
-    internal_datafusion_err!("{}", message.into())
-}
+// `proto_error` (returns `DataFusionError`) lives at
+// `datafusion_common::proto::proto_error` after the proto-common <->
+// datafusion-common decoupling. proto-common itself has no datafusion-*
+// deps.
